refactor the stat code
